# recipe-client/src/main/java/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: gaalruben02@gmail.com
    name: Ga�l Ruben
  description: description = "Yokudlela Recipes/Tables API for Graphical User Interface
    .
  title: Yokudlela Recipes/Menus API
  version: v1
servers:
- description: local dev
  url: http://recipe:8080/recipe
paths:
  /menu/add/product:
    post:
      operationId: addConsumer
      parameters:
      - description: Term�k neve
        explode: true
        in: query
        name: name
        required: true
        schema:
          type: string
        style: form
      - description: Mennyis�g
        explode: true
        in: query
        name: quantity
        required: true
        schema:
          format: double
          type: number
        style: form
      - description: M�rt�kegys�g
        explode: true
        in: query
        name: unit
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
          description: Sikeres felvitel
      security:
      - openid:
        - recipe
      summary: �j term�k felvitele
      tags:
      - menu-controller
      x-accepts: application/json
...
  /menu/getLowQuantityProducts:
    get:
      operationId: getLowQuantityProducts
      parameters:
      - description: Maxim�lis mennyis�g
        explode: true
        in: query
        name: quantity
        required: true
        schema:
          format: double
          type: number
        style: form
      - description: M�rt�kegys�g
        explode: true
        in: query
        name: unit
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Product'
                type: array
          description: Sikeres lek�rdez�s
      summary: "Azon term�kek lek�rdez�se, amelyek a megadott m�rt�kegys�get (G, KG,\
        \ PCS, L, ML) haszn�lj�k �s a param�terk�nt �tadott mennyis�gn�l (pl. 100)\
        \ kevesebb van bel?l�k rakt�ron."
      tags:
      - menu-controller
      x-accepts: application/json
components:
  schemas:
    Product:
      description: Term�k
      example:
        unit: PCS
        quantity: 6.027456183070403
        name: name
        id: 0
      properties:
        id:
          format: int64
          type: integer
        name:
          description: Term�k neve
          type: string
        quantity:
          description: Rakt�ron l�v? mennyis�g
          format: double
          type: number
        unit:
          description: M�rt�kegys�g
          enum:
          - PCS
          - KG
          - G
          - L
          - ML
          type: string
      type: object
...
  securitySchemes:
    openid:
      description: KeyCloak Yokudlela
      openIdConnectUrl: http://localhost:6080/auth/realms/yokudlela/.well-known/openid-configuration
      type: openIdConnect